CollectIlluminaLaneMetrics doesn't require a read structure.
CollectIlluminaLaneMetrics will fail if skip or molecular barcode bases are found in the read structure.
While we could try to infer in which read (index1, read1, read2, or index2) those bases were read and
then convert them to template bases, the better solution is to not pass a read stucture at all, and
allow CollectIlluminaLaneMetrics to read the read structure from the RunInfo.xml.