Various Task Manage Speedups (#373)
* Pipeline: do not add children of tasks already added

* Pipeline: solve stack overflow in addChildren

* Task: find cycles in a collection of tasks all at once.

When a task returns a list of tasks via getTasks, we currently check for
cyclical dependencies on each task independently.  If they are all
connected in the DAG, then this is really slow!  This avoids that
by finding the strongly connected components jointly across the new *to
be added* tasks.

* Small performance optimizations on dense dags

* TaskManager: exponential reduction in sleep time if we can do anything,
otherwise linear increase

* No need to check for cycles before executing a task, as it was checked
when added.

* Warn if a single step in execution takes longer than 30 seconds

* Optimize GraphNode

* update based on review

* ResourceSet bug when subsetting resources and some performance optimizations (#374)

* ResourceSet bug fix

If the minimum to subset to is fractional, with a different fractional
value than the maximum, it could be missed.

* Task Manager optimizations

* A few NaiveScheduler simplifications