Bugfix/nested non unit tasks (#379)
* [bugfix] tasks that return an empty tasks are not executed

Currently, a task that returns an empty task (`EmptyTask` via its
`getTasks` method would be stuck in the "ready" state.  It is not a
`UnitTasks` so isn't scheduled and executed, but it is also not an
`EmptyTask` so isn't considered "running".

Treat all non-`UnitTask` tasks that have no predecessors as "running",
and therefore eligible for completion.